Fix Requests Initialization errors
